IntegrateBoardsCommandHandler
