+++ src/OctoshiftCLI.Tests/ado2gh/Commands/IntegrateBoardsCommandTests.cs
@@ -16,7 +16,7 @@ namespace OctoshiftCLI.Tests.AdoToGithub.Commands
         private readonly Mock<OctoLogger> _mockOctoLogger = TestHelpers.CreateMock<OctoLogger>();
         private readonly Mock<EnvironmentVariableProvider> _mockEnvironmentVariableProvider = TestHelpers.CreateMock<EnvironmentVariableProvider>();
 
-        private readonly IntegrateBoardsCommand _command;
+        private readonly IntegrateBoardsCommandHandler _command;
 
         private const string ADO_ORG = "FooOrg";
         private const string ADO_TEAM_PROJECT = "BlahTeamProject";
@@ -32,23 +32,24 @@ namespace OctoshiftCLI.Tests.AdoToGithub.Commands
 
         public IntegrateBoardsCommandTests()
         {
-            _command = new IntegrateBoardsCommand(_mockOctoLogger.Object, _mockAdoApiFactory.Object, _mockEnvironmentVariableProvider.Object);
+            _command = new IntegrateBoardsCommandHandler(_mockOctoLogger.Object, _mockAdoApiFactory.Object, _mockEnvironmentVariableProvider.Object);
         }
 
         [Fact]
         public void Should_Have_Options()
         {
-            Assert.NotNull(_command);
-            Assert.Equal("integrate-boards", _command.Name);
-            Assert.Equal(7, _command.Options.Count);
+            var command = new IntegrateBoardsCommand(_mockOctoLogger.Object, _mockAdoApiFactory.Object, _mockEnvironmentVariableProvider.Object);
+            Assert.NotNull(command);
+            Assert.Equal("integrate-boards", command.Name);
+            Assert.Equal(7, command.Options.Count);
 
-            TestHelpers.VerifyCommandOption(_command.Options, "ado-org", true);
-            TestHelpers.VerifyCommandOption(_command.Options, "ado-team-project", true);
-            TestHelpers.VerifyCommandOption(_command.Options, "github-org", true);
-            TestHelpers.VerifyCommandOption(_command.Options, "github-repo", true);
-            TestHelpers.VerifyCommandOption(_command.Options, "ado-pat", false);
-            TestHelpers.VerifyCommandOption(_command.Options, "github-pat", false);
-            TestHelpers.VerifyCommandOption(_command.Options, "verbose", false);
+            TestHelpers.VerifyCommandOption(command.Options, "ado-org", true);
+            TestHelpers.VerifyCommandOption(command.Options, "ado-team-project", true);
+            TestHelpers.VerifyCommandOption(command.Options, "github-org", true);
+            TestHelpers.VerifyCommandOption(command.Options, "github-repo", true);
+            TestHelpers.VerifyCommandOption(command.Options, "ado-pat", false);
+            TestHelpers.VerifyCommandOption(command.Options, "github-pat", false);
+            TestHelpers.VerifyCommandOption(command.Options, "verbose", false);
         }
 
         [Fact]
